+++ NEWS.md
@@ -1,5 +1,6 @@
 # tern 0.6.4.9000
 
+* Added `conf.ini.coxph`, `conf.ini.survfit`, `conf.ini.ztest`, `conf.type`, `probs` arguments in `t_tte`
 * Fix bug in `t_rsp` of not showing p-vlaue, odds ratio and CIs when `strata_data` is not `NULL`
 * Add stratified analysis for `t_forest_rsp` and `t_forest_tte`, stratified analysis is footnoted in `g_forest`.
 * Addes `footnotes`, `footnotes<-` and `add_footnotes<-` functions to deal with footnotes